[api] Rename authentication service to pix authentication service
