+++ api/lib/domain/usecases/index.js
@@ -7,7 +7,6 @@ const dependencies = {
   assessmentRepository: require('../../infrastructure/repositories/assessment-repository'),
   assessmentResultRepository: require('../../infrastructure/repositories/assessment-result-repository'),
   authenticationMethodRepository: require('../../infrastructure/repositories/authentication-method-repository'),
-  authenticationService: require('../../domain/services/authentication-service'),
   authenticationSessionService: require('../../domain/services/authentication/authentication-session-service'),
   badgeAcquisitionRepository: require('../../infrastructure/repositories/badge-acquisition-repository'),
   badgeCriteriaService: require('../../domain/services/badge-criteria-service'),
@@ -100,6 +99,7 @@ const dependencies = {
   participationsForCampaignManagementRepository: require('../../infrastructure/repositories/participations-for-campaign-management-repository'),
   partnerCertificationScoringRepository: require('../../infrastructure/repositories/partner-certification-scoring-repository'),
   passwordGenerator: require('../../domain/services/password-generator'),
+  pixAuthenticationService: require('../services/pix-authentication-service'),
   pickChallengeService: require('../services/pick-challenge-service'),
   pixPlusDroitMaitreCertificationResultRepository: require('../../infrastructure/repositories/pix-plus-droit-maitre-certification-result-repository'),
   pixPlusDroitExpertCertificationResultRepository: require('../../infrastructure/repositories/pix-plus-droit-expert-certification-result-repository'),
