+++ api/tests/unit/domain/services/authentication-service_test.js
@@ -6,7 +6,7 @@ const User = require('../../../../lib/domain/models/User');
 
 const encryptionService = require('../../../../lib/domain/services/encryption-service');
 
-const authenticationService = require('../../../../lib/domain/services/authentication-service');
+const pixAuthenticationService = require('../../../../lib/domain/services/pix-authentication-service');
 
 describe('Unit | Domain | Services | authentication-service', function () {
   describe('#getUserByUsernameAndPassword', function () {
@@ -39,7 +39,7 @@ describe('Unit | Domain | Services | authentication-service', function () {
 
       it('should call the user repository', async function () {
         // when
-        await authenticationService.getUserByUsernameAndPassword({
+        await pixAuthenticationService.getUserByUsernameAndPassword({
           username,
           password,
           userRepository,
@@ -54,7 +54,7 @@ describe('Unit | Domain | Services | authentication-service', function () {
         const expectedPasswordHash = authenticationMethod.authenticationComplement.password;
 
         // when
-        await authenticationService.getUserByUsernameAndPassword({
+        await pixAuthenticationService.getUserByUsernameAndPassword({
           username,
           password,
           userRepository,
@@ -69,7 +69,7 @@ describe('Unit | Domain | Services | authentication-service', function () {
 
       it('should return user found', async function () {
         // when
-        const foundUser = await authenticationService.getUserByUsernameAndPassword({
+        const foundUser = await pixAuthenticationService.getUserByUsernameAndPassword({
           username,
           password,
           userRepository,
@@ -87,7 +87,7 @@ describe('Unit | Domain | Services | authentication-service', function () {
         userRepository.getByUsernameOrEmailWithRolesAndPassword.rejects(new UserNotFoundError());
 
         // when
-        const error = await catchErr(authenticationService.getUserByUsernameAndPassword)({
+        const error = await catchErr(pixAuthenticationService.getUserByUsernameAndPassword)({
           username,
           password,
           userRepository,
@@ -103,7 +103,7 @@ describe('Unit | Domain | Services | authentication-service', function () {
         encryptionService.checkPassword.rejects(new PasswordNotMatching());
 
         // when
-        const error = await catchErr(authenticationService.getUserByUsernameAndPassword)({
+        const error = await catchErr(pixAuthenticationService.getUserByUsernameAndPassword)({
           username,
           password,
           userRepository,
