+++ api/tests/unit/domain/usecases/authenticate-user_test.js
@@ -10,7 +10,7 @@ const {
   UserShouldChangePasswordError,
 } = require('../../../../lib/domain/errors');
 
-const authenticationService = require('../../../../lib/domain/services/authentication-service');
+const pixAuthenticationService = require('../../../../lib/domain/services/pix-authentication-service');
 const endTestScreenRemovalService = require('../../../../lib/domain/services/end-test-screen-removal-service');
 const appMessages = require('../../../../lib/domain/constants');
 
@@ -30,7 +30,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
       getByUsernameOrEmailWithRoles: sinon.stub(),
       updateLastLoggedAt: sinon.stub(),
     };
-    sinon.stub(authenticationService, 'getUserByUsernameAndPassword');
+    sinon.stub(pixAuthenticationService, 'getUserByUsernameAndPassword');
     sinon.stub(endTestScreenRemovalService, 'isEndTestScreenRemovalEnabledForSomeCertificationCenter');
   });
 
@@ -42,7 +42,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
     const expirationDelaySeconds = 1;
     const user = domainBuilder.buildUser({ email: userEmail });
 
-    authenticationService.getUserByUsernameAndPassword.resolves(user);
+    pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
     refreshTokenService.createRefreshTokenFromUserId
       .withArgs({
         userId: user.id,
@@ -63,7 +63,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
     });
 
     // then
-    expect(authenticationService.getUserByUsernameAndPassword).to.have.been.calledWithExactly({
+    expect(pixAuthenticationService.getUserByUsernameAndPassword).to.have.been.calledWithExactly({
       username: userEmail,
       password,
       userRepository,
@@ -78,7 +78,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
     const expirationDelaySeconds = 1;
     const user = domainBuilder.buildUser({ email: userEmail });
 
-    authenticationService.getUserByUsernameAndPassword.resolves(user);
+    pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
     refreshTokenService.createAccessTokenFromRefreshToken.resolves({ accessToken, expirationDelaySeconds });
 
     // when
@@ -97,7 +97,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
   it('should rejects an error when given username (email) does not match an existing one', async function () {
     // given
     const unknownUserEmail = 'unknown_user_email@example.net';
-    authenticationService.getUserByUsernameAndPassword.rejects(new UserNotFoundError());
+    pixAuthenticationService.getUserByUsernameAndPassword.rejects(new UserNotFoundError());
 
     // when
     const error = await catchErr(authenticateUser)({
@@ -112,7 +112,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
 
   it('should rejects an error when given password does not match the found user’s one', async function () {
     // given
-    authenticationService.getUserByUsernameAndPassword.rejects(new MissingOrInvalidCredentialsError());
+    pixAuthenticationService.getUserByUsernameAndPassword.rejects(new MissingOrInvalidCredentialsError());
 
     // when
     const error = await catchErr(authenticateUser)({
@@ -130,7 +130,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
       // given
       const scope = appMessages.PIX_ORGA.SCOPE;
       const user = new User({ email: userEmail, memberships: [] });
-      authenticationService.getUserByUsernameAndPassword.resolves(user);
+      pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
 
       const expectedErrorMessage = appMessages.PIX_ORGA.NOT_LINKED_ORGANIZATION_MSG;
 
@@ -151,7 +151,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
       // given
       const scope = appMessages.PIX_ADMIN.SCOPE;
       const user = new User({ email: userEmail, pixAdminRoles: [] });
-      authenticationService.getUserByUsernameAndPassword.resolves(user);
+      pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
 
       const expectedErrorMessage = appMessages.PIX_ADMIN.NOT_SUPERADMIN_MSG;
 
@@ -173,7 +173,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
         // given
         const scope = appMessages.PIX_CERTIF.SCOPE;
         const user = domainBuilder.buildUser({ email: userEmail, certificationCenterMemberships: [] });
-        authenticationService.getUserByUsernameAndPassword.resolves(user);
+        pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
         endTestScreenRemovalService.isEndTestScreenRemovalEnabledForSomeCertificationCenter.resolves(false);
 
         const expectedErrorMessage = appMessages.PIX_CERTIF.NOT_LINKED_CERTIFICATION_MSG;
@@ -203,7 +203,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
         });
 
         endTestScreenRemovalService.isEndTestScreenRemovalEnabledForSomeCertificationCenter.resolves(true);
-        authenticationService.getUserByUsernameAndPassword.resolves(user);
+        pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
         refreshTokenService.createRefreshTokenFromUserId
           .withArgs({
             userId: user.id,
@@ -226,7 +226,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
         });
 
         // then
-        expect(authenticationService.getUserByUsernameAndPassword).to.have.been.calledWithExactly({
+        expect(pixAuthenticationService.getUserByUsernameAndPassword).to.have.been.calledWithExactly({
           username: userEmail,
           password,
           userRepository,
@@ -246,7 +246,7 @@ describe('Unit | Application | UseCase | authenticate-user', function () {
       });
       user.authenticationMethods = [authenticationMethod];
 
-      authenticationService.getUserByUsernameAndPassword.resolves(user);
+      pixAuthenticationService.getUserByUsernameAndPassword.resolves(user);
 
       // when
       const error = await catchErr(authenticateUser)({
